fix: file dialog defaults to Documents (#168); feature search covers copy tray, macros, abbreviations (#171)
#168: wx.FileDialog had no defaultDir, so on first launch the dialog
opened in the install directory (AppData\Local\Programs).  Added
_file_dialog_default_dir() which returns the last-used directory when
one is known, or wx.StandardPaths.GetDocumentsDir() as a fallback.
open_file and save_file_as now pass this value and record the chosen
directory in self._last_file_dir after each successful operation.

#171: "Why Don't I See a Feature?" returned no results for "copy tray",
"macros"/"macro", "abbreviations", "snippet", "watch folder", "agent",
and related terms because FEATURE_DEFINITIONS had no aliases for them.
Added aliases and short descriptions to core.edit, core.macros,
core.watch_folder, and future.ai.

GATE-11: growing features.py triggered the module-size budget.
Extracted FeatureDefinition, FEATURE_DEFINITIONS, and FEATURE_ALIASES
to feature_catalog.py (same pattern as the earlier feature_command_map.py
extraction). features.py 1029 -> 680 lines; budget updated.
